op_alloy_genesis
Module params

Module containing fee parameters.

Structs

OpBaseFeeParams

Optimism Base Fee Configuration

Constants

BASE_SEPOLIA_BASE_FEE_PARAMS

Get the base fee parameters for Base Sepolia.

BASE_SEPOLIA_EIP1559_DEFAULT_ELASTICITY_MULTIPLIER

Base fee max change denominator for Base Sepolia as defined in the Optimism transaction costs doc.

OP_MAINNET_BASE_FEE_PARAMS

Get the base fee parameters for Optimism Mainnet.

OP_MAINNET_EIP1559_BASE_FEE_MAX_CHANGE_DENOMINATOR_CANYON

Base fee max change denominator for Optimism Mainnet as defined in the Optimism Canyon hardfork.

OP_MAINNET_EIP1559_DEFAULT_BASE_FEE_MAX_CHANGE_DENOMINATOR

Base fee max change denominator for Optimism Mainnet as defined in the Optimism transaction costs doc.

OP_MAINNET_EIP1559_DEFAULT_ELASTICITY_MULTIPLIER

Base fee max change denominator for Optimism Mainnet as defined in the Optimism transaction costs doc.

OP_SEPOLIA_BASE_FEE_PARAMS

Get the base fee parameters for Optimism Sepolia.

OP_SEPOLIA_EIP1559_BASE_FEE_MAX_CHANGE_DENOMINATOR_CANYON

Base fee max change denominator for Optimism Sepolia as defined in the Optimism Canyon hardfork.

OP_SEPOLIA_EIP1559_DEFAULT_BASE_FEE_MAX_CHANGE_DENOMINATOR

Base fee max change denominator for Optimism Sepolia as defined in the Optimism transaction costs doc.

OP_SEPOLIA_EIP1559_DEFAULT_ELASTICITY_MULTIPLIER

Base fee max change denominator for Optimism Sepolia as defined in the Optimism transaction costs doc.

Functions

base_fee_params

Returns the [BaseFeeParams] for the given chain id.