op_alloy_genesis
Module system

System Config Type

Structs

SystemAccounts

System accounts

SystemConfig

System configuration.

Enums

BatcherUpdateError

An error for updating the batcher address on the SystemConfig.

EIP1559UpdateError

An error for updating the EIP-1559 parameters on the SystemConfig.

GasConfigUpdateError

An error for updating the gas config on the SystemConfig.

GasLimitUpdateError

An error for updating the gas limit on the SystemConfig.

LogProcessingError

An error occurred while processing the update log.

SystemConfigUpdateError

An error for processing the SystemConfig update log.

SystemConfigUpdateType

Represents type of update to the system config.

Constants

CONFIG_UPDATE_EVENT_VERSION_0

The initial version of the system config event log.

CONFIG_UPDATE_TOPIC

keccak256("ConfigUpdate(uint256,uint8,bytes)")