op_alloy_genesis
Module rollup

Rollup Config Types

Structs

RollupConfig

The Rollup configuration.

Constants

BASE_MAINNET_CONFIG

The RollupConfig for Base Mainnet.

BASE_SEPOLIA_CONFIG

The RollupConfig for Base Sepolia.

FJORD_MAX_SEQUENCER_DRIFT

The max sequencer drift when the Fjord hardfork is active.

GRANITE_CHANNEL_TIMEOUT

The channel timeout once the Granite hardfork is active.

MAX_RLP_BYTES_PER_CHANNEL_BEDROCK

The max rlp bytes per channel for the Bedrock hardfork.

MAX_RLP_BYTES_PER_CHANNEL_FJORD

The max rlp bytes per channel for the Fjord hardfork.

OP_MAINNET_CONFIG

The RollupConfig for OP Mainnet.

OP_SEPOLIA_CONFIG

The RollupConfig for OP Sepolia.

Functions

default_granite_channel_timeout 🔒 serde

rollup_config_from_chain_id

Returns the rollup config for the given chain ID.